+++ CopilotKit/packages/react-ui/CHANGELOG.md
@@ -1,5 +1,22 @@
 # ui
 
+## 1.9.2-next.12
+
+### Patch Changes
+
+- Updated dependencies [3f8c575]
+  - @copilotkit/react-core@1.9.2-next.12
+  - @copilotkit/runtime-client-gql@1.9.2-next.12
+  - @copilotkit/shared@1.9.2-next.12
+
+## 1.9.2-next.11
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.9.2-next.11
+- @copilotkit/react-core@1.9.2-next.11
+- @copilotkit/shared@1.9.2-next.11
+
 ## 1.9.2-next.10
 
 ### Patch Changes