+++ CopilotKit/packages/runtime/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @copilotkit/runtime
 
+## 1.9.2-next.12
+
+### Patch Changes
+
+- @copilotkit/shared@1.9.2-next.12
+
+## 1.9.2-next.11
+
+### Patch Changes
+
+- 1f4949a: - fix: remove agent discovery for agui agents
+  - @copilotkit/shared@1.9.2-next.11
+
 ## 1.9.2-next.10
 
 ### Patch Changes