+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @copilotkit/shared
 
+## 1.9.2-next.12
+
+## 1.9.2-next.11
+
 ## 1.9.2-next.10
 
 ## 1.9.2-next.9